Update Project : Seperated each python scripts
