+++ config.py
@@ -1,0 +1,6 @@
+import logging
+import re
+
+
+EMOTICON_ID_REGEX = re.compile("https://e.kakao.com/t/.+")
+LOG_LEVEL = logging.INFO
